Fix warnings noticed by clang (#720)
* Fix warnings noticed by clang

* change snprintf("%s") to strlcpy/tempstr_cpy

* more strlcpy to replace snprintf

* Disable rule of snprintf is better than strcpy

* Fix build

* Revert the changes not interesting

+++ src/pages/common/_model_loadsave.c
@@ -182,7 +182,7 @@ static int count_files(const char *dir, const char *ext, const char *match)
 static void press_cb(guiObject_t *obj, s8 press_type, const void *data)
 {
     (void)obj;
-    int selected = (long)data + 1;
+    int selected = (int)data + 1;
     if (press_type != -1)
         return;
